* Adds functionality to the merger tree file construction code:
  + Allows for cases with no traced particles;

  + Allows for creation of references to merger tree sections within the
    haloTree datasets.

+++ scripts/aux/Millennium_Trees_Grab.pl
@@ -28,6 +28,34 @@ if ( exists($arguments{"output"}) ) {
     $outputFile = "Millennium_Trees.csv";
 }
 
+# Specify the database table.
+if ( exists($arguments{"table"}) ) {
+    $table = $arguments{"table"};
+} else {
+    $table = "millimil..MPAHalo";
+}
+
+# Specify the snapshot table.
+if ( exists($arguments{"snapshotTable"}) ) {
+    $snapshotTable = $arguments{"snapshotTable"};
+} else {
+    $snapshotTable = "millimil..Snapshots";
+}
+
+# Specify the particle table.
+if ( exists($arguments{"particleTable"}) ) {
+    $particleTable = $arguments{"particleTable"};
+} else {
+    $particleTable = "MMSnapshots..MillimilSnapshots";
+}
+
+# Determine if particles should be traced.
+if ( exists($arguments{"traceParticles"}) ) {
+    $traceParticles = $arguments{"traceParticles"};
+} else {
+    $traceParticles = "yes";
+}
+
 # Specify the database URL.
 $databaseURL = "http://www.g-vo.org/MyMillennium3?action=doQuery&SQL=";
 
@@ -37,7 +65,7 @@ $getCommandBase .= " --http-user="  .$sqlUser     unless ( $sqlUser     eq "" );
 $getCommandBase .= " --http-passwd=".$sqlPassword unless ( $sqlPassword eq "" );
 
 # Build the SQL query to retrieve basic node data.
-$sqlQuery = $databaseURL."select node.treeId, node.haloId, node.descendantId, node.firstHaloInFOFgroupId, node.snapNum, node.redshift, node.m_tophat, node.np, node.x, node.y, node.z, node.velX, node.velY, node.velZ, node.spinX, node.spinY, node.spinZ, node.halfmassRadius, node.mostBoundID from millimil..MPAHalo node, millimil..MPAHalo root where root.haloId = node.treeId";
+$sqlQuery = $databaseURL."select node.treeId, node.haloId, node.descendantId, node.firstHaloInFOFgroupId, node.snapNum, node.redshift, node.m_tophat, node.np, node.x, node.y, node.z, node.velX, node.velY, node.velZ, node.spinX, node.spinY, node.spinZ, node.halfmassRadius, node.mostBoundID from ".$table." node, ".$table." root where root.haloId = node.treeId";
 # Append any required selection.
 $sqlQuery .= " and ".$selection unless ( $selection eq "" );
 # Add an order by statement.
@@ -46,24 +74,27 @@ $sqlQuery .= " order by node.treeId";
 $getCommand = $getCommandBase." \"".$sqlQuery."\" -O ".$outputFile;
 system($getCommand);
 
-# Build the SQL query to retrieve particle data for lost subhalos.
-$sqlQuery = $databaseURL."select node.descendantId, count(*) as num into countTable from millimil..MPAHalo node, millimil..MPAHalo root where root.haloId = node.treeId"; # and node.descendantId > 0 and node.descendantId >= 0";
-# Append any required selection.
-$sqlQuery .= " and ".$selection unless ( $selection eq "" );
-# Append grouping command.
-$sqlQuery .= " group by node.descendantId";
-# Add command to find nodes which are not the only antescendents.
-$sqlQuery .= "; select node.mostBoundId, node.snapNum into boundTable from countTable, millimil..MPAHalo node where countTable.num > 1 and node.descendantId = countTable.descendantId";
-# Add command to select most bound particles trajectories corresponding to these nodes.
-$sqlQuery .= "; select snap.id, times.redshift, snap.x, snap.y, snap.z, snap.vx, snap.vy, snap.vz from boundTable, MMSnapshots..MillimilSnapshots snap, millimil..Snapshots times where boundTable.mostBoundId = snap.id and boundTable.snapNum <= snap.snapnum and times.snapnum = snap.snapNum";
-# Add commands to drop the temporary tables;
-$sqlQuery .= "; drop table countTable; drop table boundTable";
-# Retrieve the data.
-$getCommand = $getCommandBase." \"".$sqlQuery."\" -O particles.tmp";
-system($getCommand);
-# Sort the output to get ordering by particle number and then by (decreasing) redshift.
-$outputFile =~ s/\.csv/_Particles.csv/;
-system("sort -g -k1,1 -k2,2r -t, -u particles.tmp > ".$outputFile);
-unlink("particles.tmp");
+# Trace particles if requested.
+if ( $traceParticles eq "yes" ) {
+    # Build the SQL query to retrieve particle data for lost subhalos.
+    $sqlQuery = $databaseURL."select node.descendantId, count(*) as num into countTable from ".$table." node, ".$table." root where root.haloId = node.treeId"; # and node.descendantId > 0 and node.descendantId >= 0";
+    # Append any required selection.
+    $sqlQuery .= " and ".$selection unless ( $selection eq "" );
+    # Append grouping command.
+    $sqlQuery .= " group by node.descendantId";
+    # Add command to find nodes which are not the only antescendents.
+    $sqlQuery .= "; select node.mostBoundId, node.snapNum into boundTable from countTable, ".$table." node where countTable.num > 1 and node.descendantId = countTable.descendantId";
+    # Add command to select most bound particles trajectories corresponding to these nodes.
+    $sqlQuery .= "; select snap.id, times.redshift, snap.x, snap.y, snap.z, snap.vx, snap.vy, snap.vz from boundTable, ".$particleTable." snap, ".$snapshotTable." times where boundTable.mostBoundId = snap.id and boundTable.snapNum <= snap.snapnum and times.snapnum = snap.snapNum";
+    # Add commands to drop the temporary tables;
+    $sqlQuery .= "; drop table countTable; drop table boundTable";
+    # Retrieve the data.
+    $getCommand = $getCommandBase." \"".$sqlQuery."\" -O particles.tmp";
+    system($getCommand);
+    # Sort the output to get ordering by particle number and then by (decreasing) redshift.
+    $outputFile =~ s/\.csv/_Particles.csv/;
+    system("sort -g -k1,1 -k2,2r -t, -u particles.tmp > ".$outputFile);
+    unlink("particles.tmp");
+}
 
 exit;
